Added for-loops (and removed a tiny bug)
Also, renamed Parser.ConsumeCommaSeparatedList(string, string) to
Parser.ConsumeCommaSeparatedValueList(string, string)

The bug in question was in this function : it would validate input
that had commas with nothing behind them except other commas or the
ending delimiter.

The syntax for for-loops is the word "for" followed by a for-loop header,
followed by a simple block.

A for-loop header is delimited by opening and closing parentheses, and is
divided in three parts, separated by a comma. Each part can contain an
optional statement. The second statement, if present, must be, after evaluation, a
boolean value (obviously didn't implement the check yet, as that partly
is for semantic analysis and partly for execution). If there is no second
statement, it will always default to true

Examples :

Here, the first statement is `var i = 0`, the second statement
is `i < arr.Length`, which is a comparison, therefore is a boolean,
and the third statement is `i++`. The three statements are each separated
by a comma.

for (var i = 0, i < arr.length, i++) {
	print(arr[i])
}

Here, there is no  statements at all, however, you still have to
separate the parts with commas. Therefore, since there's no looping condition,
this for-loop will run forever.

for (,,) { }