Implemented if/else statements
`if` statements are implemented using IfParselet and are represented
using IfNode

`else` statements are implemented using ElseParselet and are
represented using ElseNode.

You can also use else-if statements.

For example, take this input :

```
if (firstCondition)
    print("something")
else if (secondCondition)
    print("something else")
else
    print("okay last one")
```

The top node would be an IfNode with the following else statement
as a child. That else would have the if statement as a child,
and THAT if statement would have the last else statement as a child.

So, the hierarchy would look like that :

if (firstCondition)
+
|-- else
     +
     |-- if (secondCondition)
	 +
	 |-- else

From the view point of member acces, it can be viewed as such :

IfNode <-- if (firstCondition)
    .ElseNode <-- else
	.IfNode <-- if (secondCondition)
	    .ElseNode <-- else

Basically, it's like you wrote :
```
if (firstCondition)
    print("something")
else {
    if (secondCondition)
        print("something else")
    else
        print("okay last one")
}
```

It's a kind of linked list, if you will.

You can check if each node has a if/else child with the IfNode.HasElse
and ElseNode.HasIf properties.

A linked list is okay for now, but I think it would be better
if it was flattened, so except some change on that part some
time soon. You can still see a quick try at it in
IfParselet.cs, which was actually pretty fast, even with
a lot of nesting.

Also, you might have noticed that the blocks of each statement
doesn't have any bracket ('{' or '}'). That's because, by default,
Parser.ConsumeSimpleBlock can now consume one-liners as blocks.
You can disable that using the optional areOneLinersAllowed parameter.
For example, functions aren't allowed to be one-liners. But now
for loops, foreach loops, and if-else statements can all be
one-liners without any problem.

Ah, also made changes to the DotGenerator, should generate less
"noise".