Implemented (do-)while loops (and fixed bugs)
Implemented while loops and do-while loops, respectively in
WhileParselet and DoWhileParselet. Both loops are represented by
a WhileNode. The WhileNode.IsDoLoop property allows you to check
if this represents a do-while loop (true) or a simple while loop
(false). If it is true, then WhileNode.DoToken will hold a reference
to the `do` keyword that started the loop; otherwise it will be null.

I FORGOT ANOTHER OPERATOR OMG HOW ?! (it was the greater-than '>'
operator, if you're curious)

Fixed a bug that would arise when using semi-colons (although I don't
really like the "solution" I used to fix it).

Oh, also, ValueNode.NULL now has the empty string as a representation