Refactored DotGenerator.cs
Made GraphNode an IEnumerable<GraphNode>, which means you can use
collection initialization syntax. I also made Graph's and GraphNode's
fluent classes, although i'm not happy at all with the style it is
currently giving to most of the code.

Anyway, of course breaking change, but I think this is definitely
for the better.