EROOR HANDLING ! REJOICE !
*error

Yes, my friends, the time has come for us to finally be able to get
multiple, actually good errors, with nice error messages and a source
preview (it's a work in progress tho !). This is all very basic,
and I'm sure there's still a lot of bugs everywhere (like infite
loops or cascading errors for example), but it was already quite a
lot of work getting this to work and writing every exception and
handling every case of every parselet and toklet. I'll polish all this
in another commit, probably. For example, the error messages are
really basic for now, and I'd like to rewrite them to be more helpful
in the future (see : Elm error messages).

An example of what the parser will output if the source code has
multiple errors :

An error happened at position test.txt(28, 37) and originated from FunctionDeclarationParselet.Parse : test.txt(28, 37) : Unexpected delim ()) at location test.txt(28, 37) after a comma in a function's parameter list. Expected a parameter name or type.

26 | }
27 |
28 | func someFunc2(int [seed, min, max],) {
                                         ^ error starts here
29 |     return new Random(seed).Next(min, max)
30 | }

An error happened at position test.txt(40, 20) and originated from ForeachParselet.Parse : test.txt(40, 20) : Unexpected ident (otherCollection) at location test.txt(40, 20) in a foreach header. Expected the 'in' keyword.

38 | foreach (item in collection) { }
39 |
40 | foreach (otherItem otherCollection) {
                        ^ error starts here
41 |     if (collection.contains(otherItem)) break;
42 |

You are the sole judge of my humble progress, but I'd say this is pretty
damn good work !

Yes, I know no one uses this project, but I prefer to do this now
because it would have been a nightmare later on with semantic
analysis and all that if we had to do throw each time something
was wrong. Also, this is all just a passion/learning project,
so I do things that make me learn stuff. While writing error messages
might not have been the most exciting stuff ever, I love the
professional look that this gives to the parser, and I'm quite happy
with my work so far. Oh and it's easier for me to debug when
something goes wrong parsing-wise.

I also think this is just a nice feature to have, and it has actually
made me (re)consider some things. I could have added them in this
commit, but I think it's already big enough, and it would have been
lost in the noise tbh. For example, I'd like to add "TextSpan"s,
basically two Locations indicating a range of text.

Error recovery here is quite difficult because there's no specific
way to delimit statements (other than ';', but those are optional
so we don't rely on those), so "panic mode"-style recovery is tricky
at best. So I don't know how the way I did it is called, but it kinda
works for now, and I hope to be able to rely and expand on it later.

I implemented this using a Logger that captures and registers errors
and then prints them when asked, nothing fancy. I reworked the
exceptions a bit, but nothing too fancy here either. No, the main
bulk was really just going throught every file and re-writing every
exception so that they are actually (at least a little bit) useful,
and would handle not having completely valid nodes and tokens. That
also meant I had to add an IsValid property to every token and node
so that invalid tokens and nodes would be marked. By default, it is
set to true (the object is valid), but you can set it to false (the
object isn't valid) in the constructor using the optional parameter
isValid or by setting the property after the node or token has been
created.

(Yes, it took me way too much effort to make all those lines exactly
the same size (it makes for some really awkward sentences) and no, I
am not doing that again.)

While there are still throwing exceptions in the code, those are
mainly for completely invalid states (i.e. InvalidCallException)
that just can't (and shouldn't) be recovered from, because they
are fatal internal errors. That's Logger.Fatal for you.

Anyway, here's the detailed changelog :

 Consumers :
  - Line and column numbers should now be more consistent
  - Added the constant StringConsumer.EOF, defined as U+0003

 Exceptions :
  - LotusException is now the base class for all cutsom exceptions.
    It contains basic informations such as the Location where the
    error occured, as well as informations about the caller/thrower
  - Most Exceptions now have clearer/betetr messages
  - UnexepectedEOF has been renamed to UnexpectedEOFException, and
    UnexpectedValueType to UnexpectedValueTypeException
  - InvalidCallExceptions now has an overloaded constructor with
    a message parameter

 Nodes :
  - All nodes now have a new property IsValid of type bool, set to
    true by default, that indicates whether or not the current node
    is valid or not. You can specify the value for this property
    by using the optional parameter `isValid` of the constructor,
    or by setting it after the node has been created.
  - Both StatementNode.NULL and ValueNode.NULL aren't valid
  - ValueNode.NULL's Location is now `Location(0, 0, "<std>")`

 Tokens :
  - All tokens now have a new property IsValid of type bool, set to
    true by default, that indicates whether or not the current token
    is valid or not. You can specify the value for this property
    by using the optional parameter `isValid` of the constructor,
    or by setting it after the token has been created.
  - Tokens representing literals will throw if the isValid parameter
    is set to true but they are not given a valid value.

 Parselets:
  - While most parselet still have a sort of "sanity check" at the
    beginning to check that they have been called on the right token,
    they don't throw anymore when they encounter a normal parse error
    (missing characters, unexpected token, invalid name, etc...);
    instead they register the error to the Logger class, set the node
    as invalid, and try to get back to a somewhat working state

 Toklets :
  - See the 'Parselets' section, same changes
  - Fixed some issues with number token parsing

 Location :
  - Removed the Location.With method

 Parser :
  - Fixed a bug with with semicolons, they would not work after some
    statements
  - Changed Parser.ConsumeCommaSeparatedList's signature
    from :
	public ValueNode[] ConsumeCommaSeparatedValueList(
		string start,
		string end,
		out bool isValid,
		int maxItemCount = -1
	)
    to :
	public ValueNode[] ConsumeCommaSeparatedValueList(
		string start,
		string end,
		ref bool isValid,
		int expectedItemCount = -1
	)

    Because `out bool isValid` would overwrite the variable even if
    it was alredy assigned to in the caller function.

Basically, better error handling.