Analyzers are here !
...well I mean, they're called "Visitors" but still.

So, for the past 2-3 weeks, I've been working on implementing visitors,
which are special classes whose only job is to traverse and transform a
tree.

For example, you might want to check if a given ValueNode is a constant.
You know that strings, numbers and booleans are always constant; while
identifiers, object creation and function calls are never constant.
But the "constantness" of things like operations, complex strings,
and array literals depend on what's inside them. That means we have to
recursively traverses nodes to decide whether they are constant or not.
That's exactly what a visitor does : it "visits" each node in a tree
and executes a method on the node. Two really simple examples of that
can be found in NameChecker.cs and ConstantChecker.cs

To test the implementation, I've made a few Visitors, like
ValuePrinter and NodeGraphMaker, the latter replacing the
ToGraphNode() virtual method that StatementNodes previously had.

If you want to implement your own visitor, just write a class
that inherits from either TokenVisitor<T> or NodeVisitor<T> where
T is the type that your visitor returns. Then, implement the
Default(StatemenNode) and Default(ValueNode) methods so that the
compiler will stop screaming at you. Once that's done, you can start
writing Visit(*Node) methods to implement the actual functionally
of your visitor. Be sure to put your visitor in the right folder.

If you need examples, there's already a few visitors available.
Using the current visitors, you can :
	- Check if a ValueNode is a constant
	- Check if a ValueNode is a name (fully- and semi-qualified)
	- Flatten a tree into a list of nodes
	- Convert tokens into GraphNode
	- Convert nodes into GraphNode
	- Convert nodes into GraphNode and then color them if
	  depending on whether they're constant or not
	- Extract all top-level values from statements
	- Print tokens with as much fidelity to the original
	  file as possible
	- idem except with values
	- idem except with statements

Each of these functionalities are exposed through their respective
classes' methods as well as the static `ASTHelper` class, that groups
together all the visitors and exposes their "main entry" method.

I chose to use the visitor pattern despite its inherent complexity
because it has allowed me to implement a lot of functionality really
quickly and without having to change every *Node class just to
implement a simple thing (like Flattener or ConstantVisualizer).

By the way, I also made it so that you can use this as a CLI. That
means you can now access most functionalities (like the ones
introduced with visitors) just by adding a few words after
`dotnet run`. For example, if you write `dotnet run print`,
it will print the AST with as much resemblance to the original
input as possible, using StatementPrinter. If you use
`dotnet run print value`, it will only print the top-level
values (like assignements and function calls). Likewise,
if you write `dotnet run graph constant`, it will print a graph
(in the graphviz/dot language of course) of the AST where every
constant value is colored green, and every non-constant value is
colored red. For the default behaviour, you can just not use a
simple `dotnet run` or `dotnet run graph`. Check Program.cs for
more info.

And yes, all the commits before were sort of "coincidental" commmits
of this big change, because I found myself having to change a lot of
stuff and discovering some new bugs (like the BoolToken/Node one),
so that's why they're all on the same day, because I had to split
allll the changes I made into several tiny commits. It took time,
to say the least.

This has taken a bit of thoughts and time to go through, but I'm sure
there's still improvements to be made, so don't hesitate to open a PR
or an issue ;)