Introducing LocationRanges !
A LocationRange (or just range) is an object that points to an
entire region of text, contrary to a simple Location that can only
point to a single character. You can actually make a LocationRange
using two Locations. A LocationRange object holds the first and last
line it points to, the first and last column it points to, and the
distance between each.

Ranges are mainly useful with exceptions, because they can point to
a whole statement or value or whatever, so that when the error is
displayed, we can underline exactly where the error starts and stops.

Most things now use LocationRange instead of Location : Parser,
Tokenizer, Consumer, etc... Sorry for the API changes but who cares
I'm the only one using this whatever, it was hard enough to refactor,
wouldn't wanna use this dumpster fire.

I also improved (I hope) error handling a bit, by fixing small bugs
and adding edge cases.

Sorry for the short message, I'm tired. Good night.