Changed ValueExtractor (ASTHelper.ExtractValue)
...and added StatementExtractor.

ValueExtractor was really confusing before, because it was kinda
recursive but not really, and was far from consistent in its inner
workings. So I threw the recursivity away. If you really must, you
can probably do it yourself easily, and it brings a lot more
flexibility to all that, so thta you can choose and understand the
way it actually works.

Oh and also I made a StatementExtractor, basically the equivalent
of ValueExtractor... but for statements.