The text of GraphNodes is now better formated !
Also you can now check for equality between graphs without worrying
about the changing IDs of each node.

Before :
Since IDs were based on the hash codes of strings, and those hash
codes change each time you run the program (because of runtime
restrictions), you couldn't just check if the output of two runs
produced the same graph by just `diff`ing both output. Even if they
would visually have been the same-looking graph, you had no real way
to compare them, except by just... looking at it and taking a guess
(or I guess meticulously zooming in on each node in the graph and
comparing how it looks in both versions of the graph.)

Now :
Just use Graph.GetHashCode() or GraphNode.GetHashCode(). They will
return exactly the same hash code if the graph is visually the same
(isomorphic ? i think that's how it's called at least idk go read on
wikipedia if you're interested.) If that is not the case, they will
return a completely different hash code (if all goes well at least.)
Actually, thanks to that change, graph should now be *textually*
identical if they represent the same AST, but it's not really my main
goal so don't consider it a guarantee in the future.

Also, if you forgot to save the hash code of a previous graph, don't
worry ! As long as you have the original text, it should be all fine,
since it will now print the hash code at the bottom of graph. No more
going back in time with git just to be sure some unrelated feature has
unforseen consequences on parsing or tokenizing !

Btw, for this change, I create a `DeterministicHashCode` struct,
which is really similar to the System.HashCode struct except for the
fact it doesn't have a random seed, allowing me to have an
hash-combining system that is deterministic across calls and ultimately
runs.

Really a tiny change but I'm glad I made it, it's gonna make my life
wayyyyy easier lol.