Resolved a long-standing FIXME kinda annoying !
Basically, while we expected for consumers to never give any null value
(tokens have Token.NULL, nodes all have their own NULLs, etc...), the
Consumer<T> class would return default(T) as the default value, which is
just a literal null for most types, so... that's not very good :/

In fact, in old versions it would cause a NullPointerException in some
cases. Therefore, *specifically* for the Consumer<T> edge case,
I had to add a check to [Expression]Parser.ConsumeValue() that would
return `left` (i.e. the expression built thus far) if the token was
null, but honestly that just didn't feel right. So instead, I made
Consumer<T>.Default setable, and not nullable (I mean, in principle
at least, nullability is only a warning away bla bla bla), so that
now it just generally checks if it's the consumer's default (whether
consumer or tokenizer or whatever else, because default is an
IConsumer property), and if so returns `left`.