Removed StatementKind to simplify grammar
Yep, might do it for ExpressionKinds afterwards too, because it's
honestly just a headache -- StatementKind litterally has no purpose,
it is just a useless piece of code to replace strings with enums,
it was referenced twice, by the exact same file : LotusGrammar.
It was never used by any other code than that, and it was basically
useless, thus why I'm cutting out the middle man and allowing
requests for parslets to be from string to parslets directly.

The `GetStatementParslet` methods are still here in name, but they
now take in strings and Tokens. Just like the previous methods, they
do not have any safeguards, and will throw KeyNotFoundException (as
well as NullPointerException for null tokens). However, there is a
new method, `TryGetStatementParslet` that return true or false
depending on whether or not it found a corresponding parslet, and
has an out parameter to return the hypothetically found parslet.
If it didn't find anything, the out parameter will be set to null.

Note : I might do the same for ExpressionKinds later on, I haven't
really decided yet, who knows.