Fixed a bug in Consumer<T>
It seems like Consumer<T> wasn't update when adding IConsumer.Default.
As a result, even though Consumer<T>.Default was properly set, it wasn't
used in the Consume() and Peek() methods, causing a mismatch when
the consumer was empty, since it just returned null (when it should
return a value that the constructor was provided, which is the value of
Default).

Check the next commit and guess how I discovered this bug :P
Definitively didn't take me 15 minutes to discover, who would do that...