Fixed bug in interpolated strings on invalid input
Take an obviously invalid input like this :

```
$"Hello {name ! Today is {Date.Today.ToString("US")} !"
	     ^
	     |
      Closing bracket
      should be here
```

This would be accepted by both the tokenizer and the parser !

To be exact, after parsing it would be stored as `$"Hello {name}" !`,
completely ignoring any text between where the end of the first
section should be and of the first **valid** section (here, the
second one). Outrageous !

I might excuse this for the tokenizer, because it is relatively
simple and cannot make too many assumptions. But this is definitely
way too big to go undetected in the parser !

It was caused by three things :
- The consume a complex string, the tokenizer just takes whatever
  token is between an opening and closing bracket, even if it's
  another opening bracket (it doesn't have any context, so it can't
  make assumptions about what that bracket could be, and just leaves
  it to later phases to decide if its valid or not), and then just
  puts it in the corresponding code section.
- The expression parser only consumes as much of the input as it can
  fit in a single value. Therefore, it only consumed `name` and then
  stopped, because in most other context it would just get consumed
  on the next round ; however, here it should always one value and
  only one
- StringParselet just trusted the expression parser to parse
  everything and detect if anything was wrong. It didn't check
  whether all tokens had been consumed or anything like that.
  However, since a code section contains *every* token between the
  opening and closing bracket, that would mean that when the section
  was invalid (i.e. there was more than one value (or more likely
  just rubish and normal text)), it would just ignore it and return
  happily

Combined, those three factors made it possible for this sneaky,
pretty specific bug to sneak in, unnoticed since probably the
addition of interpolated strings into the language/parser.