You can now query a Token's trivia
Tokens now have two HasTrivia method :
	bool HasTrivia(string, out TriviaToken)
	bool Hastrivia(TriviaKind, out TriviaToken)

They allow you to check for the existence of a trivia (both leading
or trailing), either by their representation or their kind. It
returns bool, but also has an out parameter that allows you to have
a reference to the original trivia, and thus more info.