Changed some Peek() calls into Consume(out Token)
Should be *slightly* cleaner and more performant.

Also, SimpleBlock is now a record.

PS: Please never do that. It was a stupid decision that I only continued
because I was too lazy to revert the weird changes I made,
and I'm pretty sure there was also a lot of hidden bugs before this that
*should* be cleared with this rewrite. But that has sapped my sanity and
motivation to work on this project for multiple months, because it's a
nightmare to transform.