Big logger rewrite !
TLDR: The logger will now use the errors' properties, and has a new
display style. It's also easier to work with :)

As a consequence of the last commit, the logging system is now more
extensible, flexible and (imo) slightly better organized/separated.
For example, if you want to change how locations are handled and/or
displayed (which is what pushed me to finally do this whole thing I
had been putting off for quite a while), you don't have to change a
million different files and constructors to do the same action, and
you don't have to check that they're all consistent. (Although this
design also means that the logger cannot accept System.Exception as
an error type)

As the logger now has the role of decomposing and formatting errors
based on their type and interfaces, source code formatting moved out
of it and into the SourceCode class, which is a sort of "utility"
instance class that allows for more efficient code retrieving and
displaying/formatting. This requires using URI paths instead of
normal string, which seems like a completely fine compromise, and yet
another good excuse to "secure" further the code from invalid states.

However, a downside of the current string-based logging design is
that color in error messages is practically impossible. Thus, the
next logging rework (which shouldn't be in too long anyway) will
probably be the first to include a dependency, most probably
Spectre.Console, as it is a very capable library with a large range
of very useful features (although I'll have to change some things
around, as the current "markup" method that the library uses would be
a problem as we're displaying code & stuff, therefore I'll most
probably use some sort of wrapper that will use types to represent
markup info and will escape color sequences before Spectre.Console
processes them).

                             --- PS ---
    Truth be told, I am not happy with this commit or the current
    state of error handling/logging in this project rn, but I have
    been working on it for too long now, and this is already big
    enough, so I'll just do the rest of the work in another commit