Markup styles won't 'erase' each other anymore
Styles (as well as colors and formats, obviously) are now managed
using a stack during rendering, meaning that when you stop using a
style, it will restore the one that was previously used. For example,
printing color text after having .PushStyle'ed will only affect the
color of *that* text, and will use the right style (including color)
for all the content that came before a potential .PopStyle().

PS: It doesn't matter if all the styling wasn't popped before calling
    render, or even if there were too many that were popped ;)