Added strongly-typed names !
Added a new NameNode abstract class, from which both IdentNode and
FullNameNode inherit. A NameNode is made up of multiple "parts",
which are just IdentTokens. Its ctor also requires a Token, for
both performance and i-couldnt-be-bothered reasons. For IdentNode,
this will simply be the IdentToken it encapsulates; for FullNameNode,
it's

A lot of nodes now use NameNode, both for safety and practicality,
since it's far easier and faster to extract each parts of a name
when it's from an IdentToken[] rather than an OperationNode tree.

I also added a new error, NotANameError, although it's not widely
used right now, since it's mostly a relic of an earlier prototype of
this commit. Currently it has no special handling by Logger, and is
just treated as UnexpectedError<ValueNode>.

I also plan on using Consume<T> more, probably after the current wip
commit is done (structs or changes to funcdef parsing).