Refactor tuples and their parsing.
TupleNodes are now just a wrapper around a base TupleNode<ValueNode>,
a base class which is both IEnumerable<T> and a ValueNode.

ExpressionParser.ConsumeTuple now has an overload that takes a the
function that should be used to parse each tuple item, allowing for
a lot more flexibility (cf: next commits).