Replace all calls to Logger.Fatal to assertions
Debug.Assert is only call when building a debug build, so that elides
quite a few checks that should definitely all pass in a release
build.

Also removed BoolToken's and NumberToken's string based ctor, as
that would put the burden of convertion and checking on the token
instead of the toklets.