Add a generic Tuple<T> and refactor based on that
Added a new Tuple<T> type, which is _not_ a Node, but can be used
whenever you have some list of Ts delimited by tokens. It does not
put any constraints on T, unlike the old TupleNode<T> one. From a
model-semantic view, it also makes more sense not to use a "node"
just for storing data that is unrelated to the syntactic element
its name references.

Tuple<T> is a record type, mainly so that it gets value-equality [1]
and other niceties, but its properties are all get-only (except for
Location and IsValid, of course).

This also means that, e.g., SimpleBlock and ParameterList are now
useless, since their functionality is equivalent to Tuple<StatementNode>
and Tuple<TParam>, respectively. Thus they've been removed.

SimpleBlock.IsOneLiner has been replaced by Utilities.IsOneLiner()

Expression.ConsumeTuple also now returns Tuple<T>s, and not
TupleNode<T>s like it used to. Since it is trivial to create a
TupleNode from a Tuple<ValueNode>, that is left to the caller to do
if necessary.

[1] However, .Items is still compared by its default eq comparer,
    which is most likely by reference.