feat(logger): improve the logger's source code fetching
The logger now has a set of "source code providers" which it queries
whenever it needs to display a localized error. Each provider just
implements a simple ISourceCodeProvider interface:

	public interface ISourceCodeProvider {
    		string Filename { get; }
    		SourceCode Source { get; }
	}

You can register a new provider at any time using Logger.RegisterSourceProvider.
When it needs source code for a file, the logger will then search
its list of providers for the one that matches the file it needs to
print.

Concretely, this means that whenever you create a StringConsumer for
a new source, you can register it, and then each error that came from
a "client" of this consumer will print the correct text seamlessly.
That's exactly what Program.GetConsumerForFile(FileInfo) does when
it needs to make a consumer from stdin, which allows Logger to print
the actual source code even when it doesn't come directly from a file.

As a bonus, caching and wrappers largely reduce logging's memory and
IO usage.