refactor: consolidate language properties into LotusFacts class
This removes the concept of "Grammar" and "ReadOnlyGrammar," which
had been getting feature creep while making more bloated. LotusFacts
is now responsible for handling everything that grammars did before
and more, except it is static, thus you don't have to pass it around
everywhere or make it into a singleton for no reason. It is supposed
to be a sort of queryable entity for language properties, whether it
is existing keywords, operator precedence, or valid access levels.

Let's face it, parsex was never gonna magically turn into an actual
parsing library, and it's been *years* since that idea has been
abandonned. While the fact that ReadOnlyGrammar survived so long is
nothing short of a miracle, it's time to get rid of it and its
omnipresence in the code base. This is not an "alternative to parser
combinators" anymore, this is a compiler for lotus, so it should
behave and be written like one. Please bear with me while i'm making all
those changes. It's taking some time but it feels good to finally
do a good ol' full-blocking Gen2 Collection in here.

There's still a few concepts that are so fundamental to the codebase
right now thata they are hard to factor out, tho. One of those are
"parslets." This is already far from an atomic commit, so I wasn't
planning on removing them in *this* commit, but I did experiment
around a bit, and I do have to say that parslets are pretty practical
with the current "architecture" of my code. Hell, removing them might
make LotusFacts mostly useless for parsing. Only the future will tell
I guess. See you then :)