feat(ast): added single item ctor to Tuple<T>
This required adding a 'T : ILocalized' constraint to T, which right
now isn't really a problem. However, if it ever becomes one, one way
to fix this would be to mark `.Location` as a required property, and
then add `[SetsRequiredMembers]` to the collection ctor. That way, a
call to the single-item ctor would have to also specify the Location
manually, but not the normal one with the tokens.