refact(sem): improve scope combining
we spend a little more time when combining scopes so we can "optimize"
the resulting CombinedScope (or, in some cases an empty or normal
scope).

Among other things, this flattens scope hierarchies, making it much
easier to debug scopes. It also removes ignore empty scopes, and in
case it only needs to "combine" a single non-empty scope, it will
return the original (non-empty) scope, relieving the user from having
to do that optimization themselves.